# Conflicts:
#	cmd/worker/main.go

+++ activities/files.go
@@ -2,10 +2,13 @@ package activities
 
 import (
 	"context"
+	"fmt"
 	"io"
 	"os"
 	"path/filepath"
+	"time"
 
+	"github.com/ansel1/merry/v2"
 	"github.com/bcc-code/bccm-flows/paths"
 	"github.com/samber/lo"
 	"go.temporal.io/sdk/activity"
@@ -13,6 +16,10 @@ import (
 
 type FileInput struct {
 	Path paths.Path
+}
+type DeletePathInput struct {
+	Path      paths.Path
+	RemoveAll bool
 }
 
 type FileResult struct {
@@ -184,10 +191,59 @@ func ListFiles(ctx context.Context, input FileInput) ([]paths.Path, error) {
 	}), err
 }
 
-func DeletePath(ctx context.Context, input FileInput) error {
+func DeletePath(ctx context.Context, input DeletePathInput) error {
 	log := activity.GetLogger(ctx)
 	activity.RecordHeartbeat(ctx, "DeletePath")
 	log.Info("Starting DeletePathActivity")
 
-	return os.RemoveAll(input.Path.Local())
+	if (input.Path.Path == "/") || (input.Path.Path == "") {
+		return merry.New("cannot delete root")
+	}
+
+	if input.RemoveAll {
+		return os.RemoveAll(input.Path.Local())
+	}
+
+	return os.Remove(input.Path.Local())
+}
+
+// WaitForFile waits until a file stops growing
+// Useful for waiting for a file to be fully uploaded, e.g. watch folders
+// Returns true if file is fully uploaded, false if failed, e.g. file doesnt exist after 5 minutes.
+func WaitForFile(ctx context.Context, input FileInput) (bool, error) {
+	logger := activity.GetLogger(ctx)
+	logger.Info("Starting WaitForFile")
+
+	// Use to cancel if file doesnt exist still after 5 minutes
+	startedAt := time.Now()
+
+	lastKnownSize := int64(0)
+	iterationsWhereSizeIsFreezed := 0
+
+	for {
+		res, err := os.Stat(input.Path.Local())
+		activity.RecordHeartbeat(ctx, res)
+		if err != nil {
+			if time.Since(startedAt) > time.Minute*5 {
+				return false, err
+			}
+			time.Sleep(time.Second * 5)
+			continue
+		}
+
+		if res.Size() < lastKnownSize {
+			return false, fmt.Errorf("file size decreased")
+		} else if res.Size() > lastKnownSize {
+			lastKnownSize = res.Size()
+			iterationsWhereSizeIsFreezed = 0
+			time.Sleep(time.Second * 5)
+			continue
+		}
+
+		iterationsWhereSizeIsFreezed++
+		if iterationsWhereSizeIsFreezed > 12 {
+			return true, nil
+		}
+		time.Sleep(time.Second * 5)
+	}
 }
